output_cli_test/requirements.txt
google-generativeai>=0.7.1
python-dotenv>=1.0.0
pyyaml